+++ example/example.dart
@@ -4,6 +4,9 @@ import 'package:objd/core.dart';
 import './packs/examplePack.dart';
 
 void main(List<String> args) {
+
+  print(gsonEncode(gsonEncode({"test":{"value":1.8,"str":"hi","arr":[10]}})));
+
   createProject(
     Project(
       name: "folder name",
