--- folder name/data/minecraft/tags/functions/load.json
@@ -1,1 +1,0 @@
-{"values":["mypack:load"]}
