--- folder name/data/minecraft/tags/functions/tick.json
@@ -1,1 +1,0 @@
-{"values":["mypack:main"]}
