--- folder name/data/mypack/functions/main.mcfunction
@@ -1,1 +1,0 @@
-execute anchored eyes positioned ^ ^ ^ anchored feet run function mypack:objd/ray1
