--- folder name/data/mypack/functions/objd/rayhit1.mcfunction
@@ -1,2 +1,0 @@
-# This is the raycast result from your raycast widget in main
-say Hi
