--- folder name/objd.json
@@ -1,1 +1,0 @@
-{"name":"folder name","path":"./","packs":[{"name":"mypack","files":{"main":["execute anchored eyes positioned ^ ^ ^ anchored feet run function mypack:objd/ray1"],"load":[],"objd/ray1":["# Generated ray1 Group Widget from main","say Hi","execute unless block ~ ~ ~ minecraft:air run tag @s add objd_ray_hit","execute unless entity @s[tag=objd_ray_hit] positioned ^ ^ ^0.25 run function mypack:objd/ray1","execute if entity @s[tag=objd_ray_hit] run function mypack:objd/rayhit1","execute if entity @s[tag=objd_ray_hit] run tag @s remove objd_ray_hit"],"objd/rayhit1":["# This is the raycast result from your raycast widget in main","say Hi"]},"main":"main","load":"load"}],"description":"This is a datapack generated with objd by Stevertus"}
