--- folder name/pack.mcmeta
@@ -1,1 +1,0 @@
-{"pack":{"pack_format":1,"description":"This is a datapack generated with objd by Stevertus"}}
