+++ lib/core.dart
@@ -8,4 +8,3 @@ export 'basic/widgets.dart';
 export 'modules/modules.dart';
 export 'wrappers/widgets.dart';
 export 'utils/widgets.dart';
-export 'entities/widgets.dart';
