--- lib/entities/widgets.dart
@@ -1,2 +1,0 @@
-export 'living_entity.dart';
-export 'player.dart';
